requirements.txt
# Core backend
fastapi~=0.115
uvicorn[standard]~=0.30
python-dotenv~=1.0
pydantic~=2.9
requests~=2.32

# Data
pandas~=2.2
numpy~=2.1
scikit-learn~=1.5

# ML / inference (portable)
transformers>=4.43
accelerate>=0.31
sentence-transformers~=3.0

# Pinecone
pinecone~=7.3

# GPU-only extra; excluded on macOS
bitsandbytes; platform_system != "Darwin"

# NOTE: install torch separately per platform (see Dockerfiles and local commands)
